templates/fastapi-server/requirements.txt
-i https://pypi.org/simple
alembic==1.15.1
amqp==5.3.1 ; python_version >= '3.6'
annotated-types==0.7.0 ; python_version >= '3.8'
anyio==4.9.0 ; python_version >= '3.9'
async-timeout==5.0.1 ; python_full_version < '3.11.3'
bcrypt==4.3.0
billiard==4.2.1 ; python_version >= '3.7'
celery[redis]==5.4.0
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
click==8.1.8 ; python_version >= '3.7'
click-didyoumean==0.3.1 ; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0 ; python_version >= '3.6'
colorama==0.4.6 ; platform_system == 'Windows'
cryptography==44.0.2
dnspython==2.7.0 ; python_version >= '3.9'
email-validator==2.2.0
exceptiongroup==1.2.2 ; python_version < '3.11'
fastapi==0.115.11
graphql-core==3.2.6 ; python_version >= '3.9' and python_version < '4.0'
greenlet==3.1.1 ; python_version < '3.14' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
idna==3.10 ; python_version >= '3.6'
kombu==5.5.0 ; python_version >= '3.8'
mako==1.3.9 ; python_version >= '3.8'
markupsafe==3.0.2 ; python_version >= '3.9'
packaging==24.2 ; python_version >= '3.8'
passlib[bcrypt]==1.7.4
prompt-toolkit==3.0.50 ; python_full_version >= '3.8.0'
psycopg2-binary==2.9.10
pycparser==2.22 ; python_version >= '3.8'
pydantic[email]==2.10.6
pydantic-core==2.27.2 ; python_version >= '3.8'
pydantic-settings==2.8.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1 ; python_version >= '3.8'
python-multipart==0.0.20
redis==5.2.1
six==1.17.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1 ; python_version >= '3.7'
sqlalchemy==2.0.39
starlette==0.46.1 ; python_version >= '3.9'
strawberry-graphql[fastapi]==0.262.5
typing-extensions==4.12.2 ; python_version >= '3.8'
tzdata==2025.1 ; python_version >= '2'
uvicorn==0.34.0
vine==5.1.0 ; python_version >= '3.6'
wcwidth==0.2.13
werkzeug==3.1.3
